requirements.txt
flask
gunicorn
ultralytics
Pillow

--find-links https://download.pytorch.org/whl/torch_stable.html
torch==2.3.0+cpu
torchvision==0.18.0+cpu
requests